requirements.txt
openai
langchain
langchain-openai
python-dotenv
colorama
langchain-groq
langchain.prompts